zephyr-panic: Move panic handler into crate
Currently the zephyr crate defines a panic handler implementation,
forcing all downstream users into that specific implementation. Ideally
users should be able to use the zephyr crate as a library while still
having a choice in which panic handler is used. So, the panic handler
has been moved into a separate zephyr-panic crate allowing users to
opt-in.
