+++ docgen/src/lib.rs
@@ -5,9 +5,8 @@
 
 use zephyr::printkln;
 
-// Reference the Zephyr crate so that the panic handler gets used.  This is only needed if no
-// symbols from the crate are directly used.
-extern crate zephyr;
+// Bring in the Zephyr panic handler.
+use zephyr_panic as _;
 
 #[no_mangle]
 extern "C" fn rust_main() {
