+++ samples/bench/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["staticlib"]
 
 [dependencies]
 zephyr = { version = "0.1.0", features = ["time-driver", "executor-zephyr"] }
+zephyr-panic = "0.1.0"
 critical-section = "1.1.2"
 heapless = "0.8"
 static_cell = "2.1"
