+++ samples/blinky/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["staticlib"]
 
 [dependencies]
 zephyr = "0.1.0"
+zephyr-panic = "0.1.0"
 log = "0.4.22"
 
 [build-dependencies]
