+++ samples/embassy/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["staticlib"]
 
 [dependencies]
 zephyr = { version = "0.1.0", features = ["time-driver"] }
+zephyr-panic = "0.1.0"
 log = "0.4.22"
 static_cell = "2.1"
 heapless = "0.8"
