+++ samples/hello_world/Cargo.toml
@@ -14,4 +14,5 @@ crate-type = ["staticlib"]
 
 [dependencies]
 zephyr = "0.1.0"
+zephyr-panic = "0.1.0"
 log = "0.4.22"
