+++ samples/hello_world/src/lib.rs
@@ -5,6 +5,8 @@
 
 use log::info;
 
+use zephyr_panic as _;
+
 #[no_mangle]
 extern "C" fn rust_main() {
     // SAFETY: `rust_main` runs once during application startup before any rust tasks
