+++ samples/philosophers/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["staticlib"]
 
 [dependencies]
 zephyr = "0.1.0"
+zephyr-panic = "0.1.0"
 
 # Dependencies that are used by build.rs.
 [build-dependencies]
